+++ src/main/java/kmg/tool/application/service/AccessorCreationService.java
@@ -1,11 +1,13 @@
 package kmg.tool.application.service;
 
+import kmg.tool.domain.service.Two2OneService;
+
 /**
  * アクセサ作成サービスインタフェース<br>
  *
  * @author KenichiroArai
  */
-public interface AccessorCreationService extends DynamicTemplateConversionService {
+public interface AccessorCreationService extends Two2OneService {
 
     // 処理なし
