+++ src/main/java/kmg/tool/application/service/impl/AccessorCreationServiceImpl.java
@@ -1,24 +1,9 @@
 package kmg.tool.application.service.impl;
 
-import java.io.BufferedReader;
-import java.io.BufferedWriter;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import kmg.core.infrastructure.type.KmgString;
-import kmg.core.infrastructure.utils.KmgPathUtils;
 import kmg.tool.application.service.AbstractInputCsvTemplateOutputProcessorService;
 import kmg.tool.application.service.AccessorCreationService;
-import kmg.tool.application.service.DynamicTemplateConversionService;
-import kmg.tool.domain.types.KmgToolGenMessageTypes;
 import kmg.tool.infrastructure.exception.KmgToolException;
 
 /**
@@ -29,102 +14,6 @@ import kmg.tool.infrastructure.exception.KmgToolException;
 @Service
 public class AccessorCreationServiceImpl extends AbstractInputCsvTemplateOutputProcessorService
     implements AccessorCreationService {
-
-    /** 入力ファイルパス */
-    private Path inputPath;
-
-    /** テンプレートファイルパス */
-    private Path templatePath;
-
-    /** 出力ファイルパス */
-    private Path outputPath;
-
-    /** テンプレートの動的変換サービス */
-    @Autowired
-    private DynamicTemplateConversionService dynamicTemplateConversionService;
-
-    /**
-     * 入力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 入力ファイルパス
-     */
-    @Override
-    public Path getInputPath() {
-
-        final Path result = this.inputPath;
-        return result;
-
-    }
-
-    /**
-     * 出力ファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return 出力ファイルパス
-     */
-    @Override
-    public Path getOutputPath() {
-
-        final Path result = this.outputPath;
-        return result;
-
-    }
-
-    /**
-     * テンプレートファイルパスを返す<br>
-     *
-     * @author KenichiroArai
-     *
-     * @sine 1.0.0
-     *
-     * @version 1.0.0
-     *
-     * @return テンプレートファイルパス
-     */
-    @Override
-    public Path getTemplatePath() {
-
-        final Path result = this.templatePath;
-        return result;
-
-    }
-
-    /**
-     * 初期化する
-     *
-     * @return true：成功、false：失敗
-     *
-     * @param inputPath
-     *                     入力ファイルパス
-     * @param templatePath
-     *                     テンプレートファイルパス
-     * @param outputPath
-     *                     出力ファイルパス
-     */
-    @SuppressWarnings("hiding")
-    @Override
-    public boolean initialize(final Path inputPath, final Path templatePath, final Path outputPath) {
-
-        final boolean result = true;
-
-        this.inputPath = inputPath;
-        this.templatePath = templatePath;
-        this.outputPath = outputPath;
-
-        return result;
-
-    }
 
     /**
      * 処理する
@@ -137,82 +26,7 @@ public class AccessorCreationServiceImpl extends AbstractInputCsvTemplateOutputP
     @Override
     public boolean process() throws KmgToolException {
 
-        boolean result = false;
-
-        /* 入力から出力の処理 */
-        final List<List<String>> csv = new ArrayList<>();
-
-        Path csvPath = null;
-
-        /* 一時ファイルの作成 */
-        try {
-
-            final String fileNameOnly = KmgPathUtils.getFileNameOnly(this.getInputPath());
-            csvPath = Files.createTempFile(fileNameOnly, "Temp.csv");
-            csvPath.toFile().deleteOnExit();
-
-        } catch (final IOException e) {
-
-            // TODO KenichiroArai 2025/03/06 メッセージ
-            final KmgToolGenMessageTypes msgType = KmgToolGenMessageTypes.NONE;
-            throw new KmgToolException(msgType, e);
-
-        }
-
-        /* 入力ファイルからCSV形式に変換してCSVファイルに出力する */
-        try (final BufferedReader brInput = Files.newBufferedReader(this.getInputPath());
-            final BufferedWriter brOutput = Files.newBufferedWriter(csvPath);) {
-
-            String line = null;
-
-            while ((line = brInput.readLine()) != null) {
-
-                line = line.replace("final", KmgString.EMPTY);
-                line = line.replace("static", KmgString.EMPTY);
-                final Pattern patternComment = Pattern.compile("/\\*\\* (\\S+)");
-                final Matcher matcherComment = patternComment.matcher(line);
-
-                if (matcherComment.find()) {
-
-                    final List<String> csvLine = new ArrayList<>();
-                    csvLine.add(matcherComment.group(1));
-                    csv.add(csvLine);
-
-                    continue;
-
-                }
-
-                final Pattern patternSrc = Pattern.compile("private\\s+((\\w|\\[\\]|<|>)+)\\s+(\\w+);");
-                final Matcher matcherSrc = patternSrc.matcher(line);
-
-                if (!matcherSrc.find()) {
-
-                    continue;
-
-                }
-
-                final List<String> csvLine = csv.getLast();
-                csvLine.add(matcherSrc.group(1));
-                csvLine.add(matcherSrc.group(3));
-                csvLine.add(matcherSrc.group(3).substring(0, 1).toUpperCase() + matcherSrc.group(3).substring(1));
-
-                brOutput.write(String.join(",", csvLine));
-                brOutput.write(System.lineSeparator());
-
-            }
-
-            result = true;
-
-        } catch (final IOException e) {
-
-            // TODO KenichiroArai 2025/03/06 メッセージ
-            final KmgToolGenMessageTypes msgType = KmgToolGenMessageTypes.NONE;
-            throw new KmgToolException(msgType, e);
-
-        }
-
-        this.dynamicTemplateConversionService.initialize(csvPath, this.templatePath, this.outputPath);
-        this.dynamicTemplateConversionService.process();
+        final boolean result = super.process();
 
         return result;
